rename localization_builder to locor.
