+++ example/lib/src/l10n/app_localizations.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:localization_builder/annotations.dart';
+import 'package:locor/annotations.dart';
 import 'messages_all.dart';
 part 'app_localizations.g.dart';
