+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
 name: example
-description: localization_builder usage example
+description: locor usage example
 version: 0.0.1
 author:
 homepage:
@@ -15,17 +15,17 @@ dev_dependencies:
   build: ^2.0.0
   intl_generator: ^0.1.0
   intl: ^0.17.0
-  localization_builder:
+  locor:
     git:
       url: https://github.com/digitalrmdy/localization-builder.git
-      path: localization_builder
+      path: locor
   build_runner: ^2.0.0
   flutter_test:
     sdk: flutter
 
 dependency_overrides:
-  localization_builder:
-    path: ../localization_builder
+  locor:
+    path: ../locor
 
 
 # For information on the generic Dart part of this file, see the
