--- localization_builder/README.md
@@ -1,2 +1,0 @@
-# localization-builder
-A builder to generate a Flutter localization file from a yml file
