+++ locor/LICENSE
