+++ locor/README.md
@@ -1,0 +1,2 @@
+# locor
+Generate a Flutter localization file from a yml file.
