+++ locor/lib/locor.dart
@@ -1,0 +1,8 @@
+library locor;
+
+import 'package:build/build.dart';
+import 'src/generators/app_localizations_generator.dart';
+import 'package:source_gen/source_gen.dart';
+
+Builder locor(BuilderOptions builderOptions) =>
+    SharedPartBuilder(const [AppLocalizationsGenerator()], 'locor');
