+++ locor/lib/src/exceptions/exceptions.dart
