+++ locor/lib/src/generators/app_localizations_generator.dart
