+++ locor/lib/src/stringsmetadatabuilder/strings_builder.dart
