+++ locor/lib/src/util/generator_utils.dart
