+++ locor/lib/src/util/string_utils.dart
