+++ locor/test/resources/strings.yaml
