+++ locor/test/string_value_args_matcher_test.dart
@@ -1,16 +1,13 @@
-
-
 import 'package:flutter_test/flutter_test.dart';
-import 'package:localization_builder/src/stringsmetadatabuilder/strings_builder.dart';
+import 'package:locor/src/stringsmetadatabuilder/strings_builder.dart';
 
-
-void main(){
-
+void main() {
   test("test StringValueArgsMatcher", () {
-  final input = '''hello,\$5 \$a \$surName''';
+    final input = '''hello,\$5 \$a \$surName''';
     final matches = StringValueArgsMatcher().findMatches(input);
     print(matches);
-    print(input.replaceAllMapped(RegExp(r"\$(\d+)"), (m)=>"\\\$${m.group(1)}"));
+    print(
+        input.replaceAllMapped(RegExp(r"\$(\d+)"), (m) => "\\\$${m.group(1)}"));
     expect(matches.length, 2);
     //TODO proper testing
   });
